+++ priv/repo/migrations/20260703000001_create_array_records.exs
@@ -1,0 +1,13 @@
+defmodule BlinkTest.Repo.Migrations.CreateArrayRecords do
+  use Ecto.Migration
+
+  def change do
+    create table(:array_records, primary_key: false) do
+      add :id, :integer, primary_key: true
+      add :ints, {:array, :integer}
+      add :strings, {:array, :string}
+      add :docs, {:array, :map}
+      add :matrix, {:array, {:array, :integer}}
+    end
+  end
+end